[#7] 리뷰 반영
