+++ src/main/java/com/hogwarts/sns/SnsApplication.java
@@ -2,7 +2,9 @@ package com.hogwarts.sns;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
 
+@EnableJpaAuditing
 @SpringBootApplication
 public class SnsApplication {
