--- src/main/java/com/hogwarts/sns/common/exception/FileUploadException.java
@@ -1,16 +1,0 @@
-package com.hogwarts.sns.common.exception;
-
-public class FileUploadException extends RuntimeException {
-
-	public FileUploadException() {
-	}
-
-	public FileUploadException(Throwable t) {
-		super(t);
-	}
-
-	public FileUploadException(String msg, Throwable cause) {
-		super(msg, cause);
-	}
-
-}
