+++ src/main/java/com/hogwarts/sns/domain/request/CreatePostRequest.java
@@ -1,0 +1,15 @@
+package com.hogwarts.sns.domain.request;
+
+import java.util.List;
+
+import org.springframework.web.multipart.MultipartFile;
+
+import lombok.Getter;
+import lombok.Setter;
+
+@Setter
+@Getter
+public class CreatePostRequest {
+	private String content;
+	private List<MultipartFile> images;
+}
