+++ src/main/java/com/hogwarts/sns/domain/request/ModifyPostRequest.java
@@ -1,0 +1,10 @@
+package com.hogwarts.sns.domain.request;
+
+import lombok.Getter;
+import lombok.Setter;
+
+@Setter
+@Getter
+public class ModifyPostRequest {
+	private String content;
+}
