+++ src/main/java/com/hogwarts/sns/domain/response/FeedResponse.java
@@ -1,0 +1,7 @@
+package com.hogwarts.sns.domain.response;
+
+import java.util.List;
+
+public class FeedResponse {
+	List<PostResponse> postResponses;
+}
