+++ src/main/java/com/hogwarts/sns/domain/response/FileInfo.java
@@ -1,0 +1,11 @@
+package com.hogwarts.sns.domain.response;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@Getter
+@AllArgsConstructor
+public class FileInfo {
+	private String fileName;
+	private String filePath;
+}
