+++ src/main/java/com/hogwarts/sns/domain/response/PostResponse.java
@@ -1,0 +1,18 @@
+package com.hogwarts.sns.domain.response;
+
+import java.io.Serializable;
+import java.util.List;
+
+import com.hogwarts.sns.domain.Image;
+import com.hogwarts.sns.domain.Post;
+
+import lombok.Builder;
+import lombok.Getter;
+
+@Builder
+@Getter
+public class PostResponse implements Serializable {
+	private Post post;
+	private List<Image> images;
+}
+
