+++ src/main/java/com/hogwarts/sns/exception/e5xx/FileUploadException.java
@@ -1,0 +1,16 @@
+package com.hogwarts.sns.exception.e5xx;
+
+public class FileUploadException extends RuntimeException {
+
+	public FileUploadException() {
+	}
+
+	public FileUploadException(Throwable t) {
+		super(t);
+	}
+
+	public FileUploadException(String msg, Throwable cause) {
+		super(msg, cause);
+	}
+
+}
