--- src/main/java/com/hogwarts/sns/image/domain/Image.java
@@ -1,55 +1,0 @@
-package com.hogwarts.sns.image.domain;
-
-import java.time.LocalDateTime;
-
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.persistence.JoinColumn;
-import javax.persistence.ManyToOne;
-
-import org.hibernate.annotations.CreationTimestamp;
-import org.hibernate.annotations.UpdateTimestamp;
-
-import com.fasterxml.jackson.annotation.JsonIgnore;
-import com.hogwarts.sns.post.domain.Post;
-
-import lombok.AccessLevel;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@AllArgsConstructor(access = AccessLevel.PRIVATE)
-@Getter
-@Entity
-public class Image {
-	@Id
-	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private long id;
-
-	private String name;
-
-	private String path;
-
-	@JsonIgnore
-	@ManyToOne
-	@JoinColumn(name = "post_id")
-	private Post post;
-
-	@CreationTimestamp
-	@Column(updatable = false, nullable = false)
-	private LocalDateTime createdAt;
-
-	@UpdateTimestamp
-	private LocalDateTime updatedAt;
-
-	public Image(Post post, String name, String path) {
-		this.post = post;
-		this.name = name;
-		this.path = path;
-	}
-
-}
