--- src/main/java/com/hogwarts/sns/image/dto/FileInfo.java
@@ -1,11 +1,0 @@
-package com.hogwarts.sns.image.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@Getter
-@AllArgsConstructor
-public class FileInfo {
-	private String fileName;
-	private String filePath;
-}
