--- src/main/java/com/hogwarts/sns/image/repository/ImageRepository.java
@@ -1,16 +1,0 @@
-package com.hogwarts.sns.image.repository;
-
-import java.util.List;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import com.hogwarts.sns.image.domain.Image;
-
-@Repository
-public interface ImageRepository extends JpaRepository<Image, Long> {
-
-	List<Image> findByPostId(Long postId);
-
-	void deleteAllByPostId(Long postId);
-}
