+++ src/main/java/com/hogwarts/sns/persistence/PostRepository.java
@@ -1,0 +1,10 @@
+package com.hogwarts.sns.persistence;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+import com.hogwarts.sns.domain.Post;
+
+@Repository
+public interface PostRepository extends JpaRepository<Post, Long> {
+}
