--- src/main/java/com/hogwarts/sns/post/dto/FeedResponse.java
@@ -1,7 +1,0 @@
-package com.hogwarts.sns.post.dto;
-
-import java.util.List;
-
-public class FeedResponse {
-	List<PostResponse> postResponses;
-}
