--- src/main/java/com/hogwarts/sns/post/dto/ModifyPostRequest.java
@@ -1,10 +1,0 @@
-package com.hogwarts.sns.post.dto;
-
-import lombok.Getter;
-import lombok.Setter;
-
-@Setter
-@Getter
-public class ModifyPostRequest {
-	private String content;
-}
