--- src/main/java/com/hogwarts/sns/user/repository/UserRepository.java
@@ -1,10 +1,0 @@
-package com.hogwarts.sns.user.repository;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import com.hogwarts.sns.user.domain.User;
-
-@Repository
-public interface UserRepository extends JpaRepository<User, Long> {
-}
