+++ src/main/resources/application-test.yml
@@ -1,6 +1,4 @@
 spring:
-  profiles:
-    active: test
   h2:
     console:
       enabled: true
